requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.txt
#
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.3.0
django-appconf==1.0.2
django-compressor==2.2
django-libsass==0.7
django-modelcluster==3.1
django-redis==3.8.2
django-taggit==0.22.1
django-treebeard==4.1.2
django==1.11.7
djangorestframework==3.6.4
elasticsearch==1.2.0
html5lib==0.999999999
idna==2.6
libsass==0.8.2
olefile==0.44
pillow==4.3.0
psycopg2==2.6
pytz==2017.3
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
unidecode==0.4.21
urllib3==1.22
uwsgi==2.0.15
wagtail==1.13.1
webencodings==0.5.1
whitenoise==3.3.1
willow==1.0
